test: rename test case
